Move the sentinel to the streaming types and keep docblocks attached
