+++ src/api/types/streaming.ts
@@ -17,3 +17,8 @@ export interface StreamCallbacks<TResult = { success: boolean; code: number }> {
    *  stop from a command error without parsing the message. */
   onConnectionLost?: () => void;
 }
+
+/** Sentinel port for the network / OTA path (vs. a server serial device path
+ *  like ``/dev/cu.usbserial-110``), carried by ``firmware/install*`` and the
+ *  ``devices/logs`` stream. */
+export const OTA_PORT = "OTA";
