+++ src/components/apply-install-method.ts
@@ -1,5 +1,5 @@
 import type { ConfiguredDevice } from "../api/types/devices.js";
-import { OTA_PORT } from "../api/types/firmware-jobs.js";
+import { OTA_PORT } from "../api/types/streaming.js";
 import type { ESPHomeFirmwareInstallDialog } from "./firmware-install-dialog.js";
 
 export interface InstallMethodHandlers {
