+++ src/components/firmware-install-dialog/install-flow.ts
@@ -3,7 +3,7 @@ import {
   JobSource,
   JobStatus,
 } from "../../api/types/firmware-jobs.js";
-import { OTA_PORT } from "../../api/types/firmware-jobs.js";
+import { OTA_PORT } from "../../api/types/streaming.js";
 import { fetchBoard } from "../../util/board-body-cache.js";
 import { chipNameToVariant, chipPlatformFamily } from "../../util/chip-variant.js";
 import { triggerDownload } from "../../util/download-text.js";
