+++ src/components/logs-dialog.ts
@@ -14,7 +14,7 @@ import {
 import { html, LitElement } from "lit";
 import { customElement, property, query, state } from "lit/decorators.js";
 import type { ESPHomeAPI } from "../api/index.js";
-import { OTA_PORT } from "../api/types/firmware-jobs.js";
+import { OTA_PORT } from "../api/types/streaming.js";
 import type { LocalizeFunc } from "../common/localize.js";
 import {
   apiConnectedContext,
