+++ src/components/logs-dialog/session.ts
@@ -1,10 +1,10 @@
-import { OTA_PORT } from "../../api/types/firmware-jobs.js";
 /**
  * The logs dialog's log source and its lifecycle.
  *
  * Every transition of ``host._session`` lives here; the element renders it and
  * owns nothing about how a stream is started, paused, or torn down.
  */
+import { OTA_PORT } from "../../api/types/streaming.js";
 import { notifyError } from "../../util/notify.js";
 import type { ESPHomeLogsDialog } from "../logs-dialog.js";
 import { isPassive, isStreaming } from "../logs-session.js";
