+++ test/components/logs-dialog-quiet-serial.test.ts
@@ -8,7 +8,7 @@ import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
 
 import { call, makeLogsDialog, session } from "./_logs-dialog-env.js";
 
-import { OTA_PORT } from "../../src/api/types/firmware-jobs.js";
+import { OTA_PORT } from "../../src/api/types/streaming.js";
 import { switchToOtaLogs } from "../../src/components/logs-dialog/session.js";
 import type { ESPHomeLogsDialog } from "./_logs-dialog-env.js";
